+++ controller/tests/utils/test_web_api_utils.py
@@ -37,8 +37,8 @@ def test_get_cloud_api_tokens__return_tokens_if_login_successful(mocker):
     assert tokens == expected_tokens
 
     mocked_post.assert_called_once_with(
-        f"https://{CLOUD_API_ENDPOINT}/users/login",
-        json={**test_creds, "client_type": f"mantarray:{CURRENT_SOFTWARE_VERSION}"},
+        f"https://{CLOUD_API_ENDPOINT}/users/login/user",
+        json={**test_creds, "service": "mantarray", "client_type": f"mantarray:{CURRENT_SOFTWARE_VERSION}"},
     )